feat: extract function split_operators into class SplitOperator, extract function check_expression into module, extract math operators and math functions into modele
