+++ final_task/operator_manager.py
@@ -1,0 +1,34 @@
+import math
+import operator
+
+
+function_dict = {
+    'abs': {'operator': abs, 'priority': 0},
+    'round': {'operator': round, 'priority': 0}
+}
+for k, v in math.__dict__.items():
+    if k.startswith('_'):
+        continue
+    function_dict[k] = {'operator': v, 'priority': 0}
+
+
+operator_dict = {
+    '+': {'operator': operator.add, 'priority': 4},
+    '-': {'operator': operator.sub, 'priority': 4},
+    '/': {'operator': operator.truediv, 'priority': 3},
+    '*': {'operator': operator.mul, 'priority': 3},
+    '%': {'operator': operator.mod, 'priority': 3},
+    '//': {'operator': operator.floordiv, 'priority': 3},
+    '^': {'operator': operator.pow, 'priority': 1},
+    '**': {'operator': operator.pow, 'priority': 1},
+    '==': {'operator': operator.eq, 'priority': 5},
+    '!=': {'operator': operator.ne, 'priority': 5},
+    '>': {'operator': operator.gt, 'priority': 5},
+    '<': {'operator': operator.lt, 'priority': 5},
+    '>=': {'operator': operator.ge, 'priority': 5},
+    '<=': {'operator': operator.le, 'priority': 5},
+}
+
+unary_dict = {
+    '-': {'operator': operator.sub, 'priority': 2}
+}
