+++ final_task/pycalc.py
@@ -4,10 +4,11 @@ import operator
 import math
 import sys
 import argparse
+from . import expression_parser, operator_manager, check_manager
 
 
 def arg_parser():
-    parser = argparse.ArgumentParser(description='Pure-python command-line calculator.', prog='pycalc')
+    parser = argparse.ArgumentParser(description='Pure-python command-line calculator.', prog='pycalc_not_my')
     parser.add_argument(
         '-m',
         '--use-modules',
@@ -19,48 +20,6 @@ def arg_parser():
     )
     expression_line = parser.parse_args().EXPRESSION
     return expression_line
-
-
-function_dict = {
-    'abs': {'operator': abs, 'priority': 0},
-    'round': {'operator': round, 'priority': 0}
-}
-for k, v in math.__dict__.items():
-    if k.startswith('_'):
-        continue
-    function_dict[k] = {'operator': v, 'priority': 0}
-
-
-operator_dict = {
-    '+': {'operator': operator.add, 'priority': 4},
-    '-': {'operator': operator.sub, 'priority': 4},
-    '/': {'operator': operator.truediv, 'priority': 3},
-    '*': {'operator': operator.mul, 'priority': 3},
-    '%': {'operator': operator.mod, 'priority': 3},
-    '//': {'operator': operator.floordiv, 'priority': 3},
-    '^': {'operator': operator.pow, 'priority': 1},
-    '**': {'operator': operator.pow, 'priority': 1},
-    '==': {'operator': operator.eq, 'priority': 5},
-    '!=': {'operator': operator.ne, 'priority': 5},
-    '>': {'operator': operator.gt, 'priority': 5},
-    '<': {'operator': operator.lt, 'priority': 5},
-    '>=': {'operator': operator.ge, 'priority': 5},
-    '<=': {'operator': operator.le, 'priority': 5},
-}
-
-unary_dict = {
-    '-': {'operator': operator.sub, 'priority': 2}
-}
-
-
-def check_expression(expression_line):
-    if not expression_line:
-        raise SyntaxError('Expression cannot be empty')
-    if expression_line.count('(') < expression_line.count(')'):
-        raise SyntaxError('Opening bracket required!')
-    elif expression_line.count('(') > expression_line.count(')'):
-        raise SyntaxError('Closing bracket required!')
-    return True
 
 
 def check_parsing_list(parsing_list):
@@ -76,109 +35,6 @@ def check_parsing_list(parsing_list):
     if parsing_list[-1] in function_dict.keys():
         raise SyntaxError('Function "{}" without argument'.format(parsing_list[-1]))
     return True
-
-
-def number_parser(number):
-    try:
-        return int(number)
-    except ValueError:
-        return float(number)
-
-
-def operator_parser(operator_symbol):
-    if operator_symbol in operator_dict.keys():
-        return operator_symbol
-    raise SyntaxError('Typo in math operator!')
-
-
-def function_parser(function_name):
-    if function_name == 'e' or function_name == 'pi':
-        return function_dict[function_name]['operator']
-    elif function_name == 'tau':
-        if sys.version_info >= (3, 6):
-            return function_dict[function_name]['operator']
-        else:
-            return 2 * function_dict['e']['operator']
-    elif function_name in function_dict.keys():
-        return function_name
-    else:
-        raise SyntaxError('There is no function with this name {}!'.format(function_name))
-
-
-def split_operators(expression_line):
-    parsing_list = []
-    last_number = ""
-    last_letter = ""
-    last_symbol = ""
-    blank_item = False
-    if check_expression(expression_line):
-        for i in expression_line:
-            if i == " ":
-                if blank_item:
-                    continue
-                blank_item = True
-                if last_symbol:
-                    parsing_list.append(operator_parser(last_symbol))
-                    last_symbol = ""
-                elif last_number:
-                    parsing_list.append(number_parser(last_number))
-                    last_number = ""
-                elif last_letter:
-                    parsing_list.append(function_parser(last_letter))
-                    last_letter = ""
-            if i.isnumeric() or i is '.':
-                if blank_item and type(parsing_list[-1]) is not str:
-                    raise SyntaxError('Blank symbol between two operands!')
-                elif blank_item:
-                    blank_item = False
-                if last_symbol:
-                    parsing_list.append(operator_parser(last_symbol))
-                    last_symbol = ""
-                if last_letter:
-                    last_letter += i
-                else:
-                    if '.' in last_number and i == '.':
-                        raise SyntaxError('Typo in the operand (two comma)!')
-                    else:
-                        last_number += i
-            elif i.isalpha():
-                if last_symbol:
-                    parsing_list.append(operator_parser(last_symbol))
-                    last_symbol = ""
-                last_letter += i
-            elif i in "!=<>/*":
-                if len(last_symbol) == 2:
-                    raise SyntaxError('Invalid operator "{}"'.format(last_symbol+i))
-                if blank_item and str(parsing_list[-1]) in '!=<>/*':
-                    raise SyntaxError('Blank symbol between twice operator')
-                elif blank_item:
-                    blank_item = False
-                if last_number:
-                    parsing_list.append(number_parser(last_number))
-                    last_number = ""
-                if last_letter:
-                    parsing_list.append(function_parser(last_letter))
-                    last_letter = ""
-                last_symbol += i
-            else:
-                if last_number:
-                    parsing_list.append(number_parser(last_number))
-                    last_number = ""
-                if last_letter:
-                    parsing_list.append(function_parser(last_letter))
-                    last_letter = ""
-                if last_symbol:
-                    parsing_list.append(operator_parser(last_symbol))
-                    last_symbol = ""
-                if i != ' ':
-                    parsing_list.append(i)
-        if last_number:
-            parsing_list.append(number_parser(last_number))
-        elif last_letter:
-            parsing_list.append(function_parser(last_letter))
-        elif last_symbol:
-            raise SyntaxError('Extra operator "{}" at the end of the expression!'.format(last_symbol))
-    return parsing_list
 
 
 def clean_add_sub_operators(last_item, converted_list):
@@ -365,7 +221,8 @@ def main():
         expression_line = arg_parser()
         operands = OperandStack()
         function = OperandStack()
-        parser = split_operators(expression_line)
+        parser = SplitOperators().split_operators
+        clear_parser = check_expression(parser)
         converted_list = converter(parser)
         result = calculate(converted_list)
         print(result)
